King Byteasar IX in Kingdom Roads
nstead, king Byteasar wants to focus on the roads, which is why he issued a new decree: from now on, all roads are to be considered cities, and all cities are to be considered roads. Now his most grateful subjects must pack up their belongings and move out from the cities to the roads, and the cartographers are busy with building a new roadRegister of the kingdom. All roads of the kingdom are to be named for the cities they connect (i.e. [city1, city2], city1 < city2), sorted lexicographically and enumerated starting from 0. A new road register for such a system needs to be created.

Your task is to help the cartographers build the new road register. Handle the challenge, and the glorious kingdom of Byteasar IX will never have to deal with its pesky cities ever again!

Example

For

roadRegister = [
  [false, true,  true,  false, false, false],
  [true,  false, false, true,  false, false],
  [true,  false, false, false, false, false],
  [false, true,  false, false, false, false],
  [false, false, false, false, false, true ],
  [false, false, false, false, true,  false]
]
the output should be

livingOnTheRoads(roadRegister) = [
  [false, true,  true,  false],
  [true,  false, false, false],
  [true,  false, false, false],
  [false, false, false, false]
]