Graphs 12 in Contours of Everything
so easy 
You decided to create an automatic image recognizer that will determine the object in the image based on its contours. The main recognition system is already implemented, and now you need to start the teaching process.

Today you want to teach your program to recognize wheel patterns, which means that you need to implement a function that, given the adjacency matrix representing the contour, will determine whether it's a wheel or not.

The wheel contour can be thought of as a single center vertex and a regular polygon with n (n > 2) vertices which are all connected to the center. Here is an example: