Graphs 13 in Contours of Everything
A book consists of a base and may have any number of pages.
The book's base consists of a single edge connecting two nodes, and it is a common edge for all the pages. Besides that, every page has only one node connected to both sides of the base.
Here is an example of a book:

Book

Given the object's contour as an undirected graph represented by its adjacency matrix adj determine whether it's a book or not.