Graphs 16  Contours of Everything
You decided to create an automatic image recognizer that will determine the object in the image based on its contours. The main recognition system is already implemented, and now you need to start the teaching process.

Today you want to teach your program to recognize flower patterns, which means that you need to implement a function that, given the adjacency matrix representing the contour, will determine whether it's a flower or not.

The flower contour consists of several (at least one) petals.
Petal contours are the same-sized (of size greater than 2) complete graphs with exactly one common vertex.
Here are some examples of the flower contours:

flowers

Given the object's contour as an undirected graph represented by adjacency matrix adj determine whether it's a flower or not.

Example

For

adj = [[false, true, true, true, true],
       [true, false, true, false, false],
       [true, true, false, false, false],
       [true, false, false, false, true],
       [true, false, false, true, false]]
the output should be
isFlower(adj) = true.