PostgreSQL 16
trungduyen0220 Easy 100 Points
Character limit: 3000 
English
Task
And the end of the year has come, and it's time for you to organize and clean up your store's disk storage. Your boss wants all types of disks to be classified and labeled with the disc's category. Each movie disc only has one main category and the contents and format of the label are as follows: CATEGORY: Name Of Film.

From the two film and category tables, return the resulting table with two columns. The first column is the label column, whose format is described as above: the genre of the film must be capitalized and the film name is formatted according to the title style, separated by colons. The second column is the description column, which is required to be converted to lowercase (not capitalized), which is the description of the movie disc. The return table needs to be sorted by label in alphabetical order.

Example
For the data of the following tables:

category

category_id	name
1	Action
2	Animation
3	Children
4	Classics
5	Comedy
film

id	title	description	category_id
1	ACADEMY DINOSAUR	A Epic Drama of a Feminist And...	1
2	ACE GOLDFINGER	A Astounding Epistle of a Data...	4
3	ADAPTATION HOLES	A Astounding Reflection of a L...	5
4	BEACH HEARTBREAKERS	A Fanciful Documentary of a Fr...	2
5	BLUES INSTINCT	A Fast-Paced Documentary of a ...	1
6	CABIN FLASH	A Intrepid Panorama of a Robot...	5
The output should be:

label	description
ACTION: Academy Dinosaur	a epic drama of a feminist and...
ACTION: Blues Instinct	a fast-paced documentary of a ...
ANIMATION: Beach Heartbreakers	a fanciful documentary of a fr...
CLASSICS: Ace Goldfinger	a astounding epistle of a data...
COMEDY: Adaptation Holes	a astounding reflection of a l...
COMEDY: Cabin Flash	a intrepid panorama of a robot...