PostgreSQL 36  Age function, INNER JOIN
In this exercise, you are going to extract a list of customers and their rental history over 90 days. You will be using the EXTRACT, DATE_TRUNC, and AGE functions and other skills that you learned to extract a data set that could be used to determine what day of the week customers are most likely to rent a DVD and the likelihood that they will return the DVD late.

The return table needs the customer_name column to store customer names, title column to store the name of the movie, rental_date to save the rental date, dayofweek column shows weekdays taken from rental_date column, The rental_days column is the period between the rental_date and return_date and the last column is the past_due column. Column past_due is TRUE when rental_days is greater than rental_duration, and FALSE otherwise.

The records need to be sorted in ascending order according to the customer_name, title and rental_date columns and the return records must be within a period of 90 days starting on 2005-05-01.

With customer table, we have structure like below:

customer_id (SMALLINT) : a surrogate primary key used to uniquely identify each customer in the table.
first_name (TEXT) : first name of customer.
last_name (TEXT) : last name of customer.
email (TEXT) : the email of customer.
active (SMALLINT) : the status of active or not of customer.
p>With film table, we have structure like below:
film_id (SMALLINT) : a surrogate primary key used to uniquely identify each film in the table.
title (TEXT) : the title of the film.
rental_duration (SMALLINT) : the length of the rental period, in days.
length (SMALLINT) :  the duration of the film, in minutes.
rating (VARCHAR[100]) : the rating assigned to the film. Can be one of: G, PG, PG-13, R, or NC-17.
With inventory table, we have structure like below:

inventory_id (SMALLINT) : a surrogate primary key used to uniquely identify each item in inventory.
film_id (SMALLINT) : a foreign key pointing to the film this item represents.
With rental table, we have structure like below:

rental_id (SMALLINT) : a surrogate primary key that uniquely identifies the rental.
rental_date (TIMESTAMP) : the date and time that the item was rented.
inventory_id (SMALLINT) : the item being rented.
return_date (TIMESTAMP) : the date and time the item was returned.